+++ packages/mantine/src/templates/GridTemplate.tsx
@@ -1,0 +1,30 @@
+import { Container, Grid } from '@mantine/core';
+import { GridTemplateProps } from '@rjsf/utils';
+
+/** Renders a `GridTemplate` for mantine, which is expecting the column sizing information coming in via the
+ * extra props provided by the caller, which are spread directly on the `Grid`/`Grid.Col`.
+ *
+ * @param props - The GridTemplateProps, including the extra props containing the Mantine grid positioning details
+ */
+export default function GridTemplate(props: GridTemplateProps) {
+  const { children, column, fluid = true, ...rest } = props;
+
+  if (column) {
+    return <Grid.Col {...rest}>{children}</Grid.Col>;
+  }
+
+  // Grid with fluid container
+  if (fluid) {
+    return (
+      <Container p='4' mx={0} w='100%'>
+        <Grid {...rest}>{children}</Grid>
+      </Container>
+    );
+  }
+  // Grid without container
+  return (
+    <Grid grow {...rest}>
+      {children}
+    </Grid>
+  );
+}