+++ packages/playground/src/components/Playground.tsx
@@ -42,6 +42,7 @@ export default function Playground({ themes, validators }: PlaygroundProps) {
     readonly: false,
     omitExtraData: false,
     liveOmit: false,
+    experimental_componentUpdateStrategy: 'customDeep',
     experimental_defaultFormStateBehavior: { arrayMinItems: 'populate', emptyObjectFields: 'populateAllDefaults' },
   });
   const [otherFormProps, setOtherFormProps] = useState<Partial<FormProps>>({});