+++ packages/utils/test/findSchemaDefinition.test.ts
@@ -1,5 +1,5 @@
-import { findSchemaDefinition, RJSFSchema } from '../src';
-import { findSchemaDefinitionRecursive } from '../src/findSchemaDefinition';
+import { findSchemaDefinition, ID_KEY, RJSFSchema } from '../src';
+import { findSchemaDefinitionRecursive, makeAllReferencesAbsolute } from '../src/findSchemaDefinition';
 
 const schema: RJSFSchema = {
   type: 'object',
@@ -46,40 +46,73 @@ const schema: RJSFSchema = {
   },
 };
 
+const internalSchema: RJSFSchema = {
+  $schema: 'https://json-schema.org/draft/2020-12/schema',
+  $id: 'https://example.com/bundled.ref.json',
+  type: 'object',
+  $defs: {
+    colors: {
+      type: 'string',
+      enum: ['red', 'green', 'blue'],
+    },
+    string: {
+      type: 'string',
+    },
+    circularRef: {
+      $ref: '/bundled.schema.json/#/$defs/circularRef',
+    },
+    undefinedRef: {
+      $ref: '#/$defs/undefined',
+    },
+  },
+  properties: {
+    num: {
+      type: 'integer',
+    },
+    string: {
+      $ref: '#/$defs/string',
+    },
+    allOf: {
+      allOf: [
+        {
+          $ref: '#/$defs/string',
+        },
+        {
+          title: 'String',
+        },
+      ],
+    },
+  },
+};
+
 const bundledSchema: RJSFSchema = {
   $schema: 'https://json-schema.org/draft/2020-12/schema',
   type: 'object',
   $id: 'https://example.com/bundled.schema.json',
   $defs: {
-    bundledSchema: {
-      $schema: 'https://json-schema.org/draft/2020-12/schema',
-      $id: 'https://example.com/bundled.ref.json',
-      type: 'object',
-      $defs: {
-        string: {
-          type: 'string',
+    bundledSchema: internalSchema,
+    bundledSchemaArray: {
+      anyOf: [
+        {
+          $schema: 'https://json-schema.org/draft/2020-12/schema',
+          $id: 'https://example.com/bundled.ref.array.1.json',
+          type: 'object',
         },
-        circularRef: {
-          $ref: '/bundled.schema.json/#/$defs/circularRef',
+        {
+          $schema: 'https://json-schema.org/draft/2020-12/schema',
+          $id: 'https://example.com/bundled.ref.array.2.json',
+          type: 'object',
         },
-        undefinedRef: {
-          $ref: '#/$defs/undefined',
-        },
-      },
-      properties: {
-        num: {
-          type: 'integer',
-        },
-        string: {
-          $ref: '#/$defs/string',
-        },
-      },
+      ],
     },
     bundledAbsoluteRef: {
       $ref: 'https://example.com/bundled.ref.json',
     },
     bundledAbsoluteRefWithAnchor: {
       $ref: 'https://example.com/bundled.ref.json/#/properties/num',
+    },
+    bundledAbsoluteRefWithinArray: {
+      $ref: 'https://example.com/bundled.ref.array.1.json',
     },
     bundledRelativeRef: {
       $ref: '/bundled.ref.json',
@@ -96,6 +129,9 @@ const bundledSchema: RJSFSchema = {
     undefinedRef: {
       $ref: '/undefined.ref.json',
     },
+    undefinedRefWithAnchor: {
+      $ref: '/bundled.ref.json#/$defs/undefinedRef',
+    },
   },
   properties: {
     undefined: {
@@ -108,15 +144,15 @@ const EXTRA_EXPECTED = { type: 'string', title: 'foo' };
 
 describe('findSchemaDefinition()', () => {
   it('throws error when ref is missing', () => {
-    expect(() => findSchemaDefinition()).toThrowError('Could not find a definition for undefined');
+    expect(() => findSchemaDefinition()).toThrow('Could not find a definition for undefined');
   });
   it('throws error when ref is malformed', () => {
-    expect(() => findSchemaDefinition('definitions/missing')).toThrowError(
+    expect(() => findSchemaDefinition('definitions/missing')).toThrow(
       'Could not find a definition for definitions/missing',
     );
   });
   it('throws error when ref does not exist', () => {
-    expect(() => findSchemaDefinition('#/definitions/missing', schema)).toThrowError(
+    expect(() => findSchemaDefinition('#/definitions/missing', schema)).toThrow(
       'Could not find a definition for #/definitions/missing',
     );
   });
@@ -130,77 +166,85 @@ describe('findSchemaDefinition()', () => {
     expect(findSchemaDefinition('#/definitions/extraNestedRef', schema)).toEqual(EXTRA_EXPECTED);
   });
   it('throws error when ref is a circular reference', () => {
-    expect(() => findSchemaDefinition('#/definitions/badCircularNestedRef', schema)).toThrowError(
+    expect(() => findSchemaDefinition('#/definitions/badCircularNestedRef', schema)).toThrow(
       'Definition for #/definitions/badCircularNestedRef is a circular reference',
     );
   });
   it('throws error when ref is a deep circular reference', () => {
-    expect(() => findSchemaDefinition('#/definitions/badCircularDeepNestedRef', schema)).toThrowError(
+    expect(() => findSchemaDefinition('#/definitions/badCircularDeepNestedRef', schema)).toThrow(
       'Definition for #/definitions/badCircularDeepNestedRef contains a circular reference through #/definitions/badCircularDeeperNestedRef -> #/definitions/badCircularDeepestNestedRef -> #/definitions/badCircularDeepNestedRef',
     );
   });
   it('throws error when bundled ref are not part of JSON Schema Draft 2020-12', () => {
-    expect(() => findSchemaDefinition('#/definitions/bundledRef', schema)).toThrowError(
+    expect(() => findSchemaDefinition('#/definitions/bundledRef', schema)).toThrow(
       'Could not find a definition for /bundled.ref.json',
     );
   });
   it('throws error when bundled ref with explicit baseURI are not part of JSON Schema Draft 2020-12', () => {
-    expect(() => findSchemaDefinition('#/properties/num', schema, 'https://example.com/bundled.ref.json')).toThrowError(
+    expect(() => findSchemaDefinition('#/properties/num', schema, 'https://example.com/bundled.ref.json')).toThrow(
       'Could not find a definition for #/properties/num',
     );
   });
   it('correctly resolves absolute bundled refs when JSON Schema Draft 2020-12', () => {
-    expect(findSchemaDefinition('#/$defs/bundledAbsoluteRef', bundledSchema)).toBe(bundledSchema.$defs!.bundledSchema);
+    expect(findSchemaDefinition('#/$defs/bundledAbsoluteRef', bundledSchema)).toStrictEqual(internalSchema);
   });
   it('correctly resolves absolute bundled refs with anchors within a JSON Schema Draft 2020-12', () => {
     expect(findSchemaDefinition('#/$defs/bundledAbsoluteRefWithAnchor', bundledSchema)).toBe(
-      (bundledSchema.$defs!.bundledSchema as RJSFSchema).properties!.num,
+      internalSchema.properties!.num,
+    );
+  });
+  it('correctly resolves absolute bundled refs in arrays within a JSON Schema Draft 2020-12', () => {
+    expect(findSchemaDefinition('#/$defs/bundledAbsoluteRefWithinArray', bundledSchema)).toBe(
+      (bundledSchema.$defs!.bundledSchemaArray as RJSFSchema).anyOf![0],
     );
   });
   it('correctly resolves absolute bundled refs within a JSON Schema Draft 2020-12 without an `$id`', () => {
     const { $id: d, ...bundledSchemaWithoutId } = bundledSchema;
-    expect(findSchemaDefinition('#/$defs/bundledAbsoluteRef', bundledSchemaWithoutId)).toBe(
-      bundledSchema.$defs!.bundledSchema,
-    );
+    expect(findSchemaDefinition('#/$defs/bundledAbsoluteRef', bundledSchemaWithoutId)).toStrictEqual(internalSchema);
   });
   it('correctly resolves relative bundled refs within a JSON Schema Draft 2020-12', () => {
-    expect(findSchemaDefinition('#/$defs/bundledRelativeRef', bundledSchema)).toBe(bundledSchema.$defs!.bundledSchema);
+    expect(findSchemaDefinition('#/$defs/bundledRelativeRef', bundledSchema)).toStrictEqual(internalSchema);
   });
   it('correctly resolves relative bundled refs with anchors within a JSON Schema Draft 2020-12', () => {
     expect(findSchemaDefinition('#/$defs/bundledRelativeRefWithAnchor', bundledSchema)).toBe(
-      (bundledSchema.$defs!.bundledSchema as RJSFSchema).$defs!.string,
+      internalSchema.$defs!.string,
     );
   });
   it('correctly resolves indirect bundled refs within a JSON Schema Draft 2020-12', () => {
-    expect(findSchemaDefinition('#/$defs/indirectRef', bundledSchema)).toBe(bundledSchema.$defs!.bundledSchema);
+    expect(findSchemaDefinition('#/$defs/indirectRef', bundledSchema)).toStrictEqual(internalSchema);
   });
   it('correctly resolves refs with explicit base URI in a bundled JSON Schema', () => {
-    expect(findSchemaDefinition('bundled.ref.json', bundledSchema, 'https://example.com/undefined.ref.json')).toBe(
-      bundledSchema.$defs!.bundledSchema,
-    );
+    expect(
+      findSchemaDefinition('bundled.ref.json', bundledSchema, 'https://example.com/undefined.ref.json'),
+    ).toStrictEqual(internalSchema);
   });
   it('correctly resolves local refs with explicit base URI in a bundled JSON Schema', () => {
     expect(findSchemaDefinition('#/properties/num', bundledSchema, 'https://example.com/bundled.ref.json')).toBe(
-      (bundledSchema.$defs!.bundledSchema as RJSFSchema).properties!.num,
+      internalSchema.properties!.num,
     );
   });
   it('throws error when relative ref is undefined in a bundled JSON Schema', () => {
-    expect(() => findSchemaDefinition('#/$defs/undefinedRef', bundledSchema)).toThrowError(
+    expect(() => findSchemaDefinition('#/$defs/undefinedRef', bundledSchema)).toThrow(
       'Could not find a definition for /undefined.ref.json',
+    );
+  });
+  it('throws error when relative ref with anchor is undefined in a bundled JSON Schema', () => {
+    expect(() => findSchemaDefinition('#/$defs/undefinedRefWithAnchor', bundledSchema)).toThrow(
+      'Could not find a definition for #/$defs/undefined',
     );
   });
   it('throws error when local ref is undefined in a bundled JSON Schema with explicit base URI', () => {
     expect(() =>
       findSchemaDefinition('#/properties/undefined', bundledSchema, 'https://example.com/bundled.ref.json'),
-    ).toThrowError('Could not find a definition for #/properties/undefined');
+    ).toThrow('Could not find a definition for #/properties/undefined');
   });
   it('throws error when explicit base URI is undefined in a bundled JSON Schema', () => {
     expect(() =>
       findSchemaDefinition('#/properties/undefined', bundledSchema, 'https://example.com/undefined.ref.json'),
-    ).toThrowError('Could not find a definition for #/properties/undefined');
+    ).toThrow('Could not find a definition for #/properties/undefined');
   });
   it('throws error when ref is a deep circular reference in a bundled JSON Schema', () => {
-    expect(() => findSchemaDefinition('#/$defs/circularRef', bundledSchema)).toThrowError(
+    expect(() => findSchemaDefinition('#/$defs/circularRef', bundledSchema)).toThrow(
       'Definition for #/$defs/circularRef contains a circular reference through /bundled.ref.json/#/$defs/circularRef -> /bundled.schema.json/#/$defs/circularRef -> #/$defs/circularRef',
     );
   });
@@ -208,15 +252,15 @@ describe('findSchemaDefinition()', () => {
 
 describe('findSchemaDefinitionRecursive()', () => {
   it('throws error when ref is missing', () => {
-    expect(() => findSchemaDefinitionRecursive()).toThrowError('Could not find a definition for undefined');
+    expect(() => findSchemaDefinitionRecursive()).toThrow('Could not find a definition for undefined');
   });
   it('throws error when ref is malformed', () => {
-    expect(() => findSchemaDefinitionRecursive('definitions/missing')).toThrowError(
+    expect(() => findSchemaDefinitionRecursive('definitions/missing')).toThrow(
       'Could not find a definition for definitions/missing',
     );
   });
   it('throws error when ref does not exist', () => {
-    expect(() => findSchemaDefinitionRecursive('#/definitions/missing', schema)).toThrowError(
+    expect(() => findSchemaDefinitionRecursive('#/definitions/missing', schema)).toThrow(
       'Could not find a definition for #/definitions/missing',
     );
   });
@@ -230,69 +274,73 @@ describe('findSchemaDefinitionRecursive()', () => {
     expect(findSchemaDefinitionRecursive('#/definitions/extraNestedRef', schema)).toEqual(EXTRA_EXPECTED);
   });
   it('throws error when ref is a circular reference', () => {
-    expect(() => findSchemaDefinitionRecursive('#/definitions/badCircularNestedRef', schema)).toThrowError(
+    expect(() => findSchemaDefinitionRecursive('#/definitions/badCircularNestedRef', schema)).toThrow(
       'Definition for #/definitions/badCircularNestedRef is a circular reference',
     );
   });
   it('throws error when ref is a deep circular reference', () => {
-    expect(() => findSchemaDefinitionRecursive('#/definitions/badCircularDeepNestedRef', schema)).toThrowError(
+    expect(() => findSchemaDefinitionRecursive('#/definitions/badCircularDeepNestedRef', schema)).toThrow(
       'Definition for #/definitions/badCircularDeepNestedRef contains a circular reference through #/definitions/badCircularDeeperNestedRef -> #/definitions/badCircularDeepestNestedRef -> #/definitions/badCircularDeepNestedRef',
     );
   });
   it('throws error when bundled ref are not part of JSON Schema Draft 2020-12', () => {
-    expect(() => findSchemaDefinitionRecursive('#/definitions/bundledRef', schema)).toThrowError(
+    expect(() => findSchemaDefinitionRecursive('#/definitions/bundledRef', schema)).toThrow(
       'Could not find a definition for /bundled.ref.json',
     );
   });
   it('throws error when bundled ref with explicit baseURI are not part of JSON Schema Draft 2020-12', () => {
     expect(() =>
       findSchemaDefinitionRecursive('#/properties/num', schema, [], 'https://example.com/bundled.ref.json'),
-    ).toThrowError('Could not find a definition for #/properties/num');
+    ).toThrow('Could not find a definition for #/properties/num');
   });
   it('correctly resolves absolute bundled refs within a JSON Schema Draft 2020-12', () => {
-    expect(findSchemaDefinitionRecursive('#/$defs/bundledAbsoluteRef', bundledSchema)).toBe(
-      bundledSchema.$defs!.bundledSchema,
-    );
+    expect(findSchemaDefinitionRecursive('#/$defs/bundledAbsoluteRef', bundledSchema)).toStrictEqual(internalSchema);
   });
   it('correctly resolves absolute bundled refs with anchors within a JSON Schema Draft 2020-12', () => {
     expect(findSchemaDefinitionRecursive('#/$defs/bundledAbsoluteRefWithAnchor', bundledSchema)).toBe(
-      (bundledSchema.$defs!.bundledSchema as RJSFSchema).properties!.num,
+      internalSchema.properties!.num,
+    );
+  });
+  it('correctly resolves absolute bundled refs in arrays within a JSON Schema Draft 2020-12', () => {
+    expect(findSchemaDefinitionRecursive('#/$defs/bundledAbsoluteRefWithinArray', bundledSchema)).toBe(
+      (bundledSchema.$defs!.bundledSchemaArray as RJSFSchema).anyOf![0],
     );
   });
   it('correctly resolves absolute bundled refs within a JSON Schema Draft 2020-12 without an `$id`', () => {
     const { $id: d, ...bundledSchemaWithoutId } = bundledSchema;
-    expect(findSchemaDefinitionRecursive('#/$defs/bundledAbsoluteRef', bundledSchemaWithoutId)).toBe(
-      bundledSchema.$defs!.bundledSchema,
+    expect(findSchemaDefinitionRecursive('#/$defs/bundledAbsoluteRef', bundledSchemaWithoutId)).toStrictEqual(
+      internalSchema,
     );
   });
   it('correctly resolves relative bundled refs within a JSON Schema Draft 2020-12', () => {
-    expect(findSchemaDefinitionRecursive('#/$defs/bundledRelativeRef', bundledSchema)).toBe(
-      bundledSchema.$defs!.bundledSchema,
-    );
+    expect(findSchemaDefinitionRecursive('#/$defs/bundledRelativeRef', bundledSchema)).toStrictEqual(internalSchema);
   });
   it('correctly resolves relative bundled refs with anchors within a JSON Schema Draft 2020-12', () => {
     expect(findSchemaDefinitionRecursive('#/$defs/bundledRelativeRefWithAnchor', bundledSchema)).toBe(
-      (bundledSchema.$defs!.bundledSchema as RJSFSchema).$defs!.string,
+      internalSchema.$defs!.string,
     );
   });
   it('correctly resolves indirect bundled refs within a JSON Schema Draft 2020-12', () => {
-    expect(findSchemaDefinitionRecursive('#/$defs/indirectRef', bundledSchema)).toBe(
-      bundledSchema.$defs!.bundledSchema,
-    );
+    expect(findSchemaDefinitionRecursive('#/$defs/indirectRef', bundledSchema)).toStrictEqual(internalSchema);
   });
   it('correctly resolves relative refs with explicit base URI in a bundled JSON Schema', () => {
     expect(
       findSchemaDefinitionRecursive('bundled.ref.json', bundledSchema, [], 'https://example.com/undefined.ref.json'),
-    ).toBe(bundledSchema.$defs!.bundledSchema);
+    ).toStrictEqual(internalSchema);
   });
   it('correctly resolves local refs with explicit base URI in a bundled JSON Schema', () => {
     expect(
       findSchemaDefinitionRecursive('#/properties/num', bundledSchema, [], 'https://example.com/bundled.ref.json'),
-    ).toBe((bundledSchema.$defs!.bundledSchema as RJSFSchema).properties!.num);
+    ).toBe(internalSchema.properties!.num);
   });
   it('throws error when relative ref is undefined in a bundled JSON Schema', () => {
-    expect(() => findSchemaDefinitionRecursive('#/$defs/undefinedRef', bundledSchema)).toThrowError(
+    expect(() => findSchemaDefinitionRecursive('#/$defs/undefinedRef', bundledSchema)).toThrow(
       'Could not find a definition for /undefined.ref.json',
+    );
+  });
+  it('throws error when relative ref with anchor is undefined in a bundled JSON Schema', () => {
+    expect(() => findSchemaDefinitionRecursive('#/$defs/undefinedRefWithAnchor', bundledSchema)).toThrow(
+      'Could not find a definition for #/$defs/undefined',
     );
   });
   it('throws error when local ref is undefined in a bundled JSON Schema with explicit base URI', () => {
@@ -303,16 +351,43 @@ describe('findSchemaDefinitionRecursive()', () => {
         [],
         'https://example.com/bundled.ref.json',
       ),
-    ).toThrowError('Could not find a definition for #/properties/undefined');
+    ).toThrow('Could not find a definition for #/properties/undefined');
   });
   it('throws error when explicit base URI is undefined in a bundled JSON Schema', () => {
     expect(() =>
       findSchemaDefinition('#/properties/undefined', bundledSchema, 'https://example.com/undefined.ref.json'),
-    ).toThrowError('Could not find a definition for #/properties/undefined');
+    ).toThrow('Could not find a definition for #/properties/undefined');
   });
   it('throws error when ref is a deep circular reference in a bundled JSON Schema', () => {
-    expect(() => findSchemaDefinitionRecursive('#/$defs/circularRef', bundledSchema, [])).toThrowError(
+    expect(() => findSchemaDefinitionRecursive('#/$defs/circularRef', bundledSchema, [])).toThrow(
       'Definition for #/$defs/circularRef contains a circular reference through /bundled.ref.json/#/$defs/circularRef -> /bundled.schema.json/#/$defs/circularRef -> #/$defs/circularRef',
     );
+  });
+});
+
+describe('makeAllReferencesAbsolute()', () => {
+  it('correctly makes all references absolute in a JSON Schema', () => {
+    expect(makeAllReferencesAbsolute(internalSchema, internalSchema[ID_KEY]!)).toStrictEqual({
+      ...internalSchema,
+      $defs: {
+        ...internalSchema.$defs,
+        circularRef: { $ref: 'https://example.com/bundled.schema.json/#/$defs/circularRef' },
+        undefinedRef: { $ref: 'https://example.com/bundled.ref.json#/$defs/undefined' },
+      },
+      properties: {
+        ...internalSchema.properties,
+        string: { $ref: 'https://example.com/bundled.ref.json#/$defs/string' },
+        allOf: {
+          allOf: [
+            {
+              $ref: 'https://example.com/bundled.ref.json#/$defs/string',
+            },
+            {
+              title: 'String',
+            },
+          ],
+        },
+      },
+    });
   });
 });